Interim continuous-adjoint shooting implementation (#885)
* ODEIntegrationComp for direct shooting is working but compute and compute_partials need to be computed

* Added new GaussLobattoGrid, RadauGrid, and UniformGrid objects to avoid muddying the line between distributions and transcriptions.

* working on fixing issues when output grid != input grid

* Removed control and polynomial control input values as IVC outputs within their respective control groups.

* Working some min time climb regression with SLSQP. Might be the new system.

* Bump minimum openmdao version to 3.19 for Vector.get_hash use.

* Some potential issues with SLSQP in the latest pyoptsparse

* Update oldest tests to use pyoptsparse 2.4.0, since 1.2 was almost 3 years old.

+++ dymos/test/test_run_problem.py
@@ -61,7 +61,7 @@ class TestRunProblem(unittest.TestCase):
 
         p.setup(check=True)
 
-        tf = np.float128(20)
+        tf = 20.0
 
         p.set_val('traj.phase0.states:x', phase0.interp('x', [1.5, 1]))
         p.set_val('traj.phase0.states:xL', phase0.interp('xL', [0, 1]))
@@ -128,7 +128,7 @@ class TestRunProblem(unittest.TestCase):
 
         p.setup(check=True)
 
-        tf = np.float128(20)
+        tf = 20.0
 
         p.set_val('traj.phase0.states:x', phase0.interp('x', [1.5, 1]))
         p.set_val('traj.phase0.states:xL', phase0.interp('xL', [0, 1]))
